requirements.txt
aioredis==1.3.1
altgraph==0.17.2
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
auto-py-to-exe==2.13.0
autobahn==22.3.2
Automat==20.2.0
backports.entry-points-selectable==1.1.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
Cython==0.29.24
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.3
dj-database-url==0.5.0
Django==4.0.4
djangorestframework==3.13.1
Eel==0.12.4
environ==1.0
filelock==3.3.1
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.12.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
jmespath==0.10.0
lxml==4.7.1
macholib==1.15.2
msgpack==1.0.3
outcome==1.1.0
packaging==21.3
parsel==1.6.0
Pillow==8.4.0
platformdirs==2.4.0
priority==1.3.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyOpenSSL==21.0.0
pyparsing==3.0.8
PySimpleGUI==4.55.1
PySocks==1.7.1
python-http-client==3.3.7
pytz==2022.1
queuelib==1.6.2
redis==4.2.2
requests==2.27.1
selenium==4.1.3
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
trio==0.19.0
trio-websocket==0.9.2
Twisted==22.1.0
txaio==22.2.1
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.9.0
w3lib==1.22.0
webdriver-manager==3.5.4
whichcraft==0.6.1
wrapt==1.14.0
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
